refactor(ui): decompose canvas_ui.py into ui/canvas package
Decomposed ui/canvas_ui.py into modular files: items.py, reference_view.py, and image_view.py. Resolved most pylint issues and updated imports in ui/app.py.
