+++ ui/canvas/__init__.py
@@ -1,0 +1,5 @@
+"""
+Canvas package.
+"""
+from ui.canvas.image_view import ImageGraphicsViewUI
+from ui.canvas.reference_view import ReferenceGraphicsViewUI
